assignment_5/requirements.txt
google-genai
python-dotenv
pydantic-settings
streamlit